Makes options programmatically added, and used.
It's all about the Option class in jokes.util.py.
It says what it's called, what the default value is, and
a function to convert strings to that type.
They are also stored in the main Dad object so that they
can be referenced more easily while making jokes, setting
the values, or getting the values. The functions for
getting and setting the values are static functions in the Joke
class.