Adds somewhat working response to user emojis.
It doesn't use all the emojis I want, and it doesn't
even recognize all of the emojis that it does use.
I have to find a different way to get the emojis besides
the unicode name, but I'm not sure what the proper way to
do that is. Perhaps a string of names according to discord, and
that list is transformed in the chore function since it has
access to the guild?

Also, it currently adds the emojis it expects for testing
purposes.

It additionally waits forever and doesn't punish incorrect
responses.