Reduces computation of favorite child.
Now the favorite child's id number is saved per guild.
Thus, when a message is received one merely needs to
check if the saved id number is the same to check if
its the favorite child.
Whenever a person's points change we recalculate the
favorite child. A bit wasteful, but this happens far
less often than a message is sent.