Big change to implement repeated "that's fair"
Two new types of OptionTypes are added.
One is NonZeroPositiveInteger, whose purpose is mysterious.
The second is "HIDDEN", and is meant to be hidden from the
list_options command. This new type is used internally to
keep track of the fair child id number, mention string, and
times left to be responded to.

The set_guild_option_value and get_guild_option methods were
changed to use guild as a parameter instead of a context object
which has a guild parameter. Not sure if it was actually required,
but it's a good change anyways.

This new joke is weird in that it also partially exists
within dad.py's on_message method. It will check if the
message author is "the fair child" of the guild, and then
respond accordingly if it is. Additionally it will not respond
to any other possible joke within that message.

Note that it is a feature that if a different user becomes
"the fair child" before the first one's time is up that
the first person gets off the hook early.