Improves the user interface with the cowsay command.
Now simply interprets the list of n words as text.
If the last word is a valid cowsay character it will
use that character. As long as the user uses punctuation
when they end a sentence with a word that happens to be
a cowsay character it will work as intended.