Upgrades image management, at least for smashing.
Images are now saved as URLS in a JSON file that is managed
by the images.py script.
This allows for Dad to not host the images, which
allows for better performance.
This would mean a loss of control over the images
remaining up though.
To address this, the images.py script also allows for
backing up all images into a folder, and subsequently
checking if those URLs are still valid.
When an image is no longer available the backup can
be posted to the appropriate hosting site (probably tenor)
and the URL modified.